+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ko.dart
@@ -947,10 +947,16 @@ class _AccessibilityTranslationsKo extends AccessibilityTranslations {
   String attachmentPositionLabel({required int index, required int total}) => '$total개 중 $index번째';
 
   @override
-  String repliedToOwnMessageLabel({required String replierName}) => '$replierName님이 내 메시지에 답장했습니다';
+  String outgoingReplyToOwnMessageLabel() => '내 메시지에 답장했습니다';
 
   @override
-  String repliedToMessageLabel({required String replierName, required String authorName}) =>
+  String outgoingReplyToMessageLabel({required String authorName}) => '$authorName님의 메시지에 답장했습니다';
+
+  @override
+  String incomingReplyToOwnMessageLabel({required String replierName}) => '$replierName님이 내 메시지에 답장했습니다';
+
+  @override
+  String incomingReplyToMessageLabel({required String replierName, required String authorName}) =>
       '$replierName님이 $authorName님의 메시지에 답장했습니다';
 
   @override